comlete feature/eventsAlbumList
